% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_yaml.R
\name{read_simple_yaml}
\alias{read_simple_yaml}
\title{Read simple yaml file}
\usage{
read_simple_yaml(filename)
}
\arguments{
\item{filename}{String, path to file}
}
\value{
A data table object indexed on key
}
\description{
Simple yaml means one "key: value" pair per line
}
