% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elements.R
\name{rotate_labels}
\alias{rotate_labels}
\title{Rotate the labels of the x axis}
\usage{
rotate_labels(angle = 60)
}
\arguments{
\item{angle}{angle by which to rotate the labels}
}
\description{
Rotate the labels of the x axis
}
