% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_yaml.R
\name{write_simple_yaml}
\alias{write_simple_yaml}
\title{Write simple yaml file}
\usage{
write_simple_yaml(filename, data)
}
\arguments{
\item{filename}{String, path to file to write to}

\item{data}{Hashmap}
}
\value{
Writes to file
}
\description{
Simple yaml means one "key: value" pair per line
}
